+++ src/components/navigation/navbar/mobile/common/BaseNavbar.vue
@@ -33,6 +33,15 @@
                 <slot></slot>
                 <li>
                     <router-link
+                        id="nav_mobile_common_course_catalog"
+                        :to="{ name: 'courseCatalog' }"
+                        class="hover:bg-blue-800 rounded-lg mobile-navbar-menu"
+                    >
+                        Course Catalog
+                    </router-link>
+                </li>
+                <li>
+                    <router-link
                         id="nav_mobile_common_about"
                         :to="{ name: 'about' }"
                         class="hover:bg-blue-800 rounded-lg mobile-navbar-menu"