+++ src/entities/AdminEntity.ts
@@ -12,6 +12,8 @@ export default class AdminEntity implements Admin {
     email = "";
     birthDate = "";
     phoneNumber = "";
+    enrollmentIdSecret = "";
+    isActive = true;
 
     constructor() {}
 }