+++ src/entities/LecturerEntity.ts
@@ -14,6 +14,8 @@ export default class LecturerEntity implements Lecturer {
     email = "";
     birthDate = "";
     phoneNumber = "";
+    enrollmentIdSecret = "";
+    isActive = true;
 
     constructor() {}
 }