+++ src/use/router/index.ts
@@ -10,12 +10,13 @@ import PageNotFound from "@/views/errors/404.vue";
 import { createRouter, createWebHistory } from "vue-router";
 const LoginView = () => import("@/views/common/Login.vue");
 const StudentCourseView = () => import("@/views/student/StudentCourseList.vue");
-const AllCourseView = () => import("@/views/shared/CourseList.vue");
+const AllCourseView = () => import("@/views/common/CourseList.vue");
 const AdminAccountListView = () => import("@/views/admin/AdminAccountList.vue");
 const CourseForm = () => import("@/views/shared/EditCreateCourseForm.vue");
 const AccountForm = () => import("@/views/admin/EditCreateAccountForm.vue");
 const PrivateProfile = () => import("@/views/common/PrivateProfile.vue");
 const PublicProfile = () => import("@/views/common/PublicProfile.vue");
+const ExamRegForm = () => import("@/views/admin/CreateExamRegForm.vue");
 const Settings = () => import("@/views/common/Settings.vue");
 const Playground = () => import("@/views/common/dev/Playground.vue");
 const SectionPlayground = () => import("@/views/common/dev/SectionPlayground.vue");
@@ -57,6 +58,15 @@ const router = createRouter({
             component: AboutPage,
             meta: {
                 title: "About" + suffix,
+            },
+        },
+        {
+            path: "/course-catalog",
+            name: "courseCatalog",
+            props: { showAllCourses: true, isCourseCatalogue: true },
+            component: AllCourseView,
+            meta: {
+                title: "Course Catalog" + suffix,
             },
         },
         {
@@ -129,6 +139,15 @@ const router = createRouter({
             meta: {
                 title: "Course Editing" + suffix,
                 roles: ["Lecturer", "Admin"],
+            },
+        },
+        {
+            path: "/create-exam-reg",
+            name: "examRegForm",
+            component: ExamRegForm,
+            meta: {
+                title: "Examination Regulation Creation" + suffix,
+                roles: ["Admin"],
             },
         },
         {
@@ -252,7 +271,7 @@ router.beforeEach(async (to, from, next) => {
 
 router.afterEach(async (to, from, next) => {
     const store = useStore();
-    store.commit(MutationTypes.FORCE_CLOSE_BURGER_MENU, true);
+    store.commit(MutationTypes.FORCE_CLOSE_BURGER_MENU);
 });
 
 export default router;