+++ tests/e2e/matriculationManagement.spec.ts
@@ -1,11 +1,13 @@
 import Admin from "@/api/api_models/user_management/Admin";
-import { FieldOfStudy } from "@/api/api_models/user_management/FieldOfStudy";
 import Student from "@/api/api_models/user_management/Student";
 import { Account } from "@/entities/Account";
 import { getMachineUserAuth, loginAsDefaultAdmin, loginAsUser, logout } from "./helpers/AuthHelper";
 import { navigateToPrivateProfile } from "./helpers/NavigationHelper";
-import { createUsers, deleteUsers, getRandomMatriculationId } from "./helpers/UserHelper";
+import { getRandomMatriculationId, createUsers, deleteUsers, getRandomizedGovernmentId } from "./helpers/UserHelper";
 import { UserWithAuth } from "./helpers/UserWithAuth";
+
+const EXAM_REG_1 = "Bachelor Computer Science v3";
+const EXAM_REG_2 = "Bachelor Computer Science v4";
 
 const random = Math.floor(Math.random() * 9999);
 let admin: Admin;
@@ -47,7 +49,8 @@ describe("Account creation, edition and deletion", function () {
                 cy.fixture("studentAuthUser.json").then((s) => {
                     (s as Account).username += random;
                     studentAuthUser = s as Account;
-                    usersWithAuth.push({ userInfo: student, auth: studentAuthUser });
+                    let governmentId = getRandomizedGovernmentId();
+                    usersWithAuth.push({ governmentId, userInfo: student, auth: studentAuthUser });
                 });
             })
             .then(async () => {
@@ -90,7 +93,7 @@ describe("Account creation, edition and deletion", function () {
         cy.get("button[id='removeFieldOfStudy-1']").should("not.exist");
         cy.get("select[id='semesterType']").select("SS");
         cy.get("select[id='semesterYear']").select("2010");
-        cy.get("select[id='fieldsOfStudy-1']").select(FieldOfStudy.COMPUTER_SCIENCE);
+        cy.get("select[id='fieldsOfStudy-1']").select(EXAM_REG_1);
         cy.get("button[id='removeFieldOfStudy-1']").click();
         cy.get("button[id='removeFieldOfStudy-1']").should("not.exist");
     });
@@ -101,7 +104,7 @@ describe("Account creation, edition and deletion", function () {
         cy.get("button[id='removeFieldOfStudy-1']").should("not.exist");
         cy.get("select[id='semesterType']").select("SS");
         cy.get("select[id='semesterYear']").select("2011");
-        cy.get("select[id='fieldsOfStudy-1']").select(FieldOfStudy.COMPUTER_SCIENCE);
+        cy.get("select[id='fieldsOfStudy-1']").select(EXAM_REG_1);
         cy.get("button[id='addImmatriculationData']").click();
         cy.wait(4000);
         cy.get("div[id='immatriculationOptions']").siblings().get("p").should("have.class", "error-message");
@@ -111,28 +114,28 @@ describe("Account creation, edition and deletion", function () {
     it("Add two fields of studies for one summer semester", function () {
         cy.get("select[id=semesterType]").select("SS");
         cy.get("select[id=semesterYear]").select("2015");
-        cy.get("select[id=fieldsOfStudy-1]").select(FieldOfStudy.COMPUTER_SCIENCE);
-        cy.get("select[id=fieldsOfStudy-2]").select(FieldOfStudy.MATHEMATICS);
+        cy.get("select[id=fieldsOfStudy-1]").select(EXAM_REG_1);
+        cy.get("select[id=fieldsOfStudy-2]").select(EXAM_REG_2);
         cy.get("button[id=addImmatriculationData]").click();
         cy.wait(5000);
     });
 
     it("Check if summer semester FoS are shown in matriculation history", function () {
         cy.get("div").contains("Immatriculation History");
-        cy.get("div").should("contain", "SS2015").and("contain", FieldOfStudy.COMPUTER_SCIENCE).and("contain", FieldOfStudy.MATHEMATICS);
+        cy.get("div").should("contain", "SS2015").and("contain", EXAM_REG_1).and("contain", EXAM_REG_2);
     });
 
     it("Add two fields of studies for one winter semester", function () {
         cy.get("select[id=semesterType]").select("WS");
         cy.get("select[id=semesterYear]").select("2015/16");
-        cy.get("select[id=fieldsOfStudy-1]").select(FieldOfStudy.COMPUTER_SCIENCE);
-        cy.get("select[id=fieldsOfStudy-2]").select(FieldOfStudy.MATHEMATICS);
+        cy.get("select[id=fieldsOfStudy-1]").select(EXAM_REG_1);
+        cy.get("select[id=fieldsOfStudy-2]").select(EXAM_REG_2);
         cy.get("button[id=addImmatriculationData]").click();
         cy.wait(5000);
     });
 
     it("Check if winter semester FoS are shown in matriculation history", function () {
-        cy.get("div").should("contain", "WS2015/16").and("contain", FieldOfStudy.COMPUTER_SCIENCE).and("contain", FieldOfStudy.MATHEMATICS);
+        cy.get("div").should("contain", "WS2015/16").and("contain", EXAM_REG_1).and("contain", EXAM_REG_2);
         logout();
     });
 
@@ -147,8 +150,8 @@ describe("Account creation, edition and deletion", function () {
     it("Check matriculation modal is filled correctly in student's profile", function () {
         cy.get("#modal-wrapper").should("exist");
         cy.get("div").contains("Immatriculation History");
-        cy.get("div").should("contain", "SS2015").and("contain", FieldOfStudy.COMPUTER_SCIENCE).and("contain", FieldOfStudy.MATHEMATICS);
-        cy.get("div").should("contain", "WS2015/16").and("contain", FieldOfStudy.COMPUTER_SCIENCE).and("contain", FieldOfStudy.MATHEMATICS);
+        cy.get("div").should("contain", "SS2015").and("contain", EXAM_REG_1).and("contain", EXAM_REG_2);
+        cy.get("div").should("contain", "WS2015/16").and("contain", EXAM_REG_1).and("contain", EXAM_REG_2);
 
         cy.get("button[id='immatriculationHistoryClose']").click();
         cy.wait(100);